+++ LethalAPI.UI.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.7.34031.279
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LethalAPI.UI", "LethalAPI.UI\LethalAPI.UI.csproj", "{5BE95D8A-21AD-4908-B27F-961258408DC4}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{5BE95D8A-21AD-4908-B27F-961258408DC4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5BE95D8A-21AD-4908-B27F-961258408DC4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5BE95D8A-21AD-4908-B27F-961258408DC4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5BE95D8A-21AD-4908-B27F-961258408DC4}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {8A77C14F-B7F7-45B4-91EF-21B06CE32623}
+	EndGlobalSection
+EndGlobal
