+++ LethalAPI.UI/Plugin.cs
@@ -1,0 +1,14 @@
+﻿using BepInEx;
+
+namespace LethalAPI.UI
+{
+	[BepInPlugin(PluginInfo.PLUGIN_GUID, PluginInfo.PLUGIN_NAME, PluginInfo.PLUGIN_VERSION)]
+	public class Plugin : BaseUnityPlugin
+	{
+		private void Awake()
+		{
+			// Plugin startup logic
+			Logger.LogInfo($"Plugin {PluginInfo.PLUGIN_GUID} is loaded!");
+		}
+	}
+}
